+++ src/frontend/src/layouts/AppLayoutHeader.vue
@@ -11,10 +11,31 @@
       </router-link>
     </div>
     <div class="header__cart">
-      <a href="cart.html">0 ₽</a>
+      <router-link to="/cart">{{ pizzaPrice }} ₽</router-link>
     </div>
-    <div class="header__user" v-if="isAuth == false">
-      <a href="#" class="header__login"><span>Войти</span></a>
+    <div class="header__user">
+      <router-link v-if="!isAuth" to="/login" class="header__login">
+        <span>Войти</span>
+      </router-link>
+      <router-link v-else to="/profile">
+        <picture>
+          <source
+            type="image/webp"
+            srcset="
+              @/assets/img/users/user5.webp    1x,
+              @/assets/img/users/user5@2x.webp 2x
+            "
+          />
+          <img
+            src="@/assets/img/users/user5.jpg"
+            srcset="@/assets/img/users/user5@2x.jpg"
+            alt="Василий Ложкин"
+            width="32"
+            height="32"
+          />
+        </picture>
+        <span>Василий Ложкин</span>
+      </router-link>
     </div>
   </header>
 </template>
@@ -28,6 +49,10 @@ export default {
       type: Boolean,
       required: true,
       default: false,
+    },
+    pizzaPrice: {
+      type: Number,
+      default: 0,
     },
   },
 };
