+++ src/frontend/src/layouts/AppLayoutSideBar.vue
@@ -1,0 +1,31 @@
+<template>
+  <div class="layout__sidebar sidebar">
+    <router-link to="/" class="logo layout__logo">
+      <img
+        src="@/assets/img/logo.svg"
+        alt="V!U!E! Pizza logo"
+        width="90"
+        height="40"
+      />
+    </router-link>
+
+    <router-link
+      class="layout__link"
+      active-class="layout__link--active"
+      to="/orders"
+      >История заказов</router-link
+    >
+    <router-link
+      class="layout__link"
+      active-class="layout__link--active"
+      to="/profile"
+      >Мои данные</router-link
+    >
+  </div>
+</template>
+
+<script>
+export default {
+  name: "AppLayoutSideBar",
+};
+</script>
