+++ src/frontend/src/views/Profile.vue
@@ -1,19 +1,5 @@
 <template>
   <main class="layout">
-    <div class="layout__sidebar sidebar">
-      <a href="index.html" class="logo layout__logo">
-        <img
-          src="@/assets/img/logo.svg"
-          alt="V!U!E! Pizza logo"
-          width="90"
-          height="40"
-        />
-      </a>
-
-      <a class="layout__link" href="#">История заказов</a>
-      <a class="layout__link layout__link--active" href="#">Мои данные</a>
-    </div>
-
     <div class="layout__content">
       <div class="layout__title">
         <h1 class="title title--big">Мои данные</h1>
